# .specify/specs/wave-0/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SENTINEL — AI-Native Engineering Operations Platform
  description: |
    Unified API for incident management, log monitoring, deployment tracking,
    AI-assisted debugging, and team collaboration.
  version: 1.0.0
  contact:
    name: SENTINEL Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.sentinel.dev/api/v1
    description: Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UUID:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        per_page:
          type: integer
          minimum: 1
          maximum: 100
        total:
          type: integer
        total_pages:
          type: integer
...
    SeverityLevel:
      type: string
      enum: [SEV1, SEV2, SEV3, SEV4]
    IncidentStatus:
      type: string
      enum: [detected, triaging, investigating, mitigating, resolved, closed]
    Incident:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        team_id:
          $ref: '#/components/schemas/UUID'
        title:
          type: string
        description:
          type: string
        severity:
          $ref: '#/components/schemas/SeverityLevel'
        status:
          $ref: '#/components/schemas/IncidentStatus'
        assigned_to:
          $ref: '#/components/schemas/UUID'
        escalated_to:
          $ref: '#/components/schemas/UUID'
        detected_at:
          $ref: '#/components/schemas/Timestamp'
        triaged_at:
          $ref: '#/components/schemas/Timestamp'
        resolved_at:
          $ref: '#/components/schemas/Timestamp'
        closed_at:
          $ref: '#/components/schemas/Timestamp'
        root_cause:
          type: string
        ai_summary:
          type: string
        ai_root_cause_ranking:
          type: array
          items:
            type: object
            properties:
              cause:
                type: string
              confidence:
                type: number
        metadata:
          type: object
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    IncidentCreate:
      type: object
      required: [title, severity]
      properties:
        title:
          type: string
          maxLength: 500
        description:
          type: string
        severity:
          $ref: '#/components/schemas/SeverityLevel'
        assigned_to:
          $ref: '#/components/schemas/UUID'
        metadata:
          type: object
...
    DeployStatus:
      type: string
      enum: [pending, in_progress, success, failed, rolled_back]
    Deployment:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        team_id:
          $ref: '#/components/schemas/UUID'
        service:
          type: string
        version:
          type: string
        environment:
          type: string
        status:
          $ref: '#/components/schemas/DeployStatus'
        deployed_by:
          $ref: '#/components/schemas/UUID'
        started_at:
          $ref: '#/components/schemas/Timestamp'
        completed_at:
          $ref: '#/components/schemas/Timestamp'
        rollback_at:
          $ref: '#/components/schemas/Timestamp'
        commit_id:
          $ref: '#/components/schemas/UUID'
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
    DeploymentCreate:
      type: object
      required: [service, version]
      properties:
        service:
          type: string
        version:
          type: string
        environment:
          type: string
          default: production
        metadata:
          type: object
    Commit:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        team_id:
          $ref: '#/components/schemas/UUID'
        deployment_id:
          $ref: '#/components/schemas/UUID'
        sha:
          type: string
        message:
          type: string
        author:
          type: string
        author_email:
          type: string
        branch:
          type: string
        repository:
          type: string
        committed_at:
          $ref: '#/components/schemas/Timestamp'
        created_at:
          $ref: '#/components/schemas/Timestamp'

    # ──────────────────────────────────────────────
    # Timeline Events
    # ──────────────────────────────────────────────
...
    TaskStatus:
      type: string
      enum: [pending, in_progress, blocked, completed, cancelled]
    Task:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        incident_id:
          $ref: '#/components/schemas/UUID'
        assigned_to:
          $ref: '#/components/schemas/UUID'
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          type: integer
          minimum: 1
          maximum: 5
        due_at:
          $ref: '#/components/schemas/Timestamp'
        completed_at:
          $ref: '#/components/schemas/Timestamp'
        completed_by:
          $ref: '#/components/schemas/UUID'
        created_at:
          $ref: '#/components/schemas/Timestamp'
        updated_at:
          $ref: '#/components/schemas/Timestamp'
...
    TaskUpdate:
      type: object
      properties:
        assigned_to:
          $ref: '#/components/schemas/UUID'
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          type: integer

    # ──────────────────────────────────────────────
    # SLA
    # ──────────────────────────────────────────────
...
    WebhookPayload:
      type: object
      properties:
        event:
          type: string
        payload:
          type: object
        source:
          type: string
security:
  - bearerAuth: []
paths:
...
  /incidents:
    get:
      tags: [incidents]
      summary: List incidents
      operationId: listIncidents
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/IncidentStatus'
        - name: severity
          in: query
          schema:
            $ref: '#/components/schemas/SeverityLevel'
        - name: assigned_to
          in: query
          schema:
            $ref: '#/components/schemas/UUID'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Incidents list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Incident'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags: [incidents]
      summary: Create incident
      operationId: createIncident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentCreate'
      responses:
        '201':
          description: Incident created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'
...
  /deployments:
    get:
      tags: [deployments]
      summary: List deployments
      operationId: listDeployments
      parameters:
        - name: service
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/DeployStatus'
        - name: environment
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Deployments list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deployment'

    post:
      tags: [deployments]
      summary: Create deployment
      operationId: createDeployment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeploymentCreate'
      responses:
        '201':
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
...
  /commits:
    get:
      tags: [commits]
      summary: List commits
      operationId: listCommits
      parameters:
        - name: repository
          in: query
          schema:
            type: string
        - name: branch
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Commits list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Commit'

  # ──────────────────────────────────────────────
  # Timeline
  # ──────────────────────────────────────────────
...
  /tasks/{task_id}:
    get:
      tags: [tasks]
      summary: Get task
      operationId: getTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

    patch:
      tags: [tasks]
      summary: Update task
      operationId: updateTask
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'

  # ──────────────────────────────────────────────
  # SLA
  # ──────────────────────────────────────────────
...
  /integrations/github/webhook:
    post:
      tags: [integrations]
      summary: GitHub webhook receiver
      operationId: githubWebhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
      responses:
        '200':
          description: Webhook processed
  /integrations/gitlab/webhook:
    post:
      tags: [integrations]
      summary: GitLab webhook receiver
      operationId: gitlabWebhook
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookPayload'
      responses:
        '200':
          description: Webhook processed

  # ──────────────────────────────────────────────
  # Realtime (SSE)
  # ──────────────────────────────────────────────
...